Master

+++ lib/big_keeper/util/cache_operator.rb
@@ -15,7 +15,9 @@ module BigKeeper
     end
 
     def load(file)
-      FileUtils.cp("#{@cache_path}/#{file}", "#{@path}/#{file}");
+      if File.exist?(@cache_path)
+        FileUtils.cp("#{@cache_path}/#{file}", "#{@path}/#{file}");
+      end
     end
 
     def clean
@@ -32,8 +34,8 @@ module BigKeeper
       FileUtils.mkdir_p(@cache_path) unless File.exist?(@cache_path)
 
       if File.exist?("#{@cache_path}/module.cache")
-        file = File.open("#{@cache_path}/module.cache", 'r')
-        @modules = JSON.load(file.read())
+        file = File.open("#{@cache_path}/module.cache", 'r', :encoding => 'UTF-8')
+        @modules = JSON.load(file.read(), :encoding => 'UTF-8')
         file.close
       else
         @modules = {"git" => {"all" => [], "current" => []}, "path" => {"all" => [], "add" => [], "del" => [], "current" => []}}
@@ -110,7 +112,7 @@ module BigKeeper
     end
 
     def cache_modules
-      file = File.new("#{@cache_path}/module.cache", 'w')
+      file = File.new("#{@cache_path}/module.cache", 'w', :encoding => 'UTF-8')
       file << @modules.to_json
       file.close
     end
